requirements.txt
pyro-ppl
torch
torchvision
